refactor builtins directory structure
merge builtin_functions/ and builtin_interfaces/ to builtins/
