+++ tests/builtins/folding/test_abs.py
@@ -3,7 +3,7 @@ from hypothesis import example, given, settings
 from hypothesis import strategies as st
 
 from vyper import ast as vy_ast
-from vyper import builtin_functions as vy_fn
+from vyper.builtins import functions as vy_fn
 from vyper.exceptions import OverflowException
