+++ tests/builtins/folding/test_addmod_mulmod.py
@@ -3,7 +3,7 @@ from hypothesis import assume, given, settings
 from hypothesis import strategies as st
 
 from vyper import ast as vy_ast
-from vyper import builtin_functions as vy_fn
+from vyper.builtins import functions as vy_fn
 
 st_uint256 = st.integers(min_value=0, max_value=2 ** 256 - 1)
