+++ tests/builtins/folding/test_epsilon.py
@@ -1,7 +1,7 @@
 import pytest
 
 from vyper import ast as vy_ast
-from vyper import builtin_functions as vy_fn
+from vyper.builtins import functions as vy_fn
 
 
 @pytest.mark.parametrize("typ_name", ["decimal"])
