+++ tests/parser/functions/test_interfaces.py
@@ -3,7 +3,7 @@ from decimal import Decimal
 import pytest
 
 from vyper.ast.signatures.interface import extract_sigs
-from vyper.builtin_interfaces import ERC20, ERC721
+from vyper.builtins.interfaces import ERC20, ERC721
 from vyper.cli.utils import extract_file_interface_imports
 from vyper.compiler import compile_code, compile_codes
 from vyper.exceptions import ArgumentException, InterfaceViolation, StructureException
