+++ vyper/ast/folding.py
@@ -2,7 +2,7 @@ import warnings
 from typing import Optional, Union
 
 from vyper.ast import nodes as vy_ast
-from vyper.builtin_functions import DISPATCH_TABLE
+from vyper.builtins.functions import DISPATCH_TABLE
 from vyper.exceptions import UnfoldableNode, UnknownType
 from vyper.semantics.types.base import VyperType
 from vyper.semantics.types.utils import type_from_annotation
