+++ vyper/ast/signatures/interface.py
@@ -3,7 +3,7 @@
 import importlib
 import pkgutil
 
-import vyper.builtin_interfaces
+import vyper.builtins.interfaces
 from vyper import ast as vy_ast
 from vyper.ast.signatures.function_signature import FunctionSignature
 from vyper.codegen.global_context import GlobalContext
@@ -12,13 +12,14 @@ from vyper.exceptions import StructureException
 
 
 # Populate built-in interfaces.
+# NOTE: code duplication with vyper/semantics/analysis/module.py
 def get_builtin_interfaces():
-    interface_names = [x.name for x in pkgutil.iter_modules(vyper.builtin_interfaces.__path__)]
+    interface_names = [x.name for x in pkgutil.iter_modules(vyper.builtins.interfaces.__path__)]
     return {
         name: extract_sigs(
             {
                 "type": "vyper",
-                "code": importlib.import_module(f"vyper.builtin_interfaces.{name}").interface_code,
+                "code": importlib.import_module(f"vyper.builtins.interfaces.{name}").interface_code,
             },
             name,
         )
