+++ vyper/builtins/_signatures.py
