+++ vyper/builtins/interfaces/ERC20Detailed.py
