+++ vyper/builtins/interfaces/ERC4626.py
