+++ vyper/builtins/interfaces/ERC721.py
