+++ vyper/semantics/README.md
@@ -52,7 +52,7 @@ Builtins are added upon initialization of the object. This includes:
 
 * Adding primitive type classes from the [`types/`](types) subpackage
 * Adding environment variables and builtin constants from [`environment.py`](environment.py)
-* Adding builtin functions from the [`functions`](../builtin_functions/functions.py) package
+* Adding builtin functions from the [`functions`](../builtins/functions.py) package
 * Adding / resetting `self` and `log`
 
 ### 2. Validating the Module Scope
